# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catering Webapp API",
        "description": "Dokumentasi API untuk sistem pemesanan paket katering (role Penjual & Pelanggan).",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi akun Pelanggan baru",
                "operationId": "ca7c0cb6772de9bffbb8c896980cebc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@example.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "no_telp": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil, langsung dapat token"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login (Penjual atau Pelanggan)",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "penjual@catering.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil, dapat token"
                    },
                    "401": {
                        "description": "Email atau password salah"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout (mencabut token yang sedang dipakai)",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "responses": {
                    "200": {
                        "description": "Berhasil logout"
                    },
                    "401": {
                        "description": "Belum login"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Data user yang sedang login",
                "operationId": "57d51f3fa328d534a6ee8562484d0687",
                "responses": {
                    "200": {
                        "description": "Data user"
                    },
                    "401": {
                        "description": "Belum login"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pelanggan/pesanan": {
            "get": {
                "tags": [
                    "Pelanggan - Pesanan"
                ],
                "summary": "Riwayat pesanan milik pelanggan yang sedang login",
                "operationId": "643fe9a98cd5b5f65db6fb96be3f4d2f",
                "responses": {
                    "200": {
                        "description": "Daftar pesanan (paginated)"
                    },
                    "401": {
                        "description": "Belum login"
                    },
                    "403": {
                        "description": "Bukan role pelanggan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pelanggan - Pesanan"
                ],
                "summary": "Bikin pesanan baru (pilih paket + lauk + gubukan + jumlah pax + tanggal acara)",
                "operationId": "d6d5ce8ae2ae48a3222701fcfa984353",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tgl_acara",
                                    "jumlah_pax",
                                    "items"
                                ],
                                "properties": {
                                    "gubukan_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "tgl_acara": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-20"
                                    },
                                    "jumlah_pax": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Tolong pedas sedang",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "paket_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "jml_paket": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "lauk_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    },
                                                    "example": [
                                                        1,
                                                        2,
                                                        3,
                                                        4,
                                                        5
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pesanan berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal atau kapasitas tanggal tersebut sudah penuh"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pelanggan/pesanan/{pesanan}": {
            "get": {
                "tags": [
                    "Pelanggan - Pesanan"
                ],
                "summary": "Detail 1 pesanan milik sendiri",
                "operationId": "639023029c33ff4f2a4e60a787503266",
                "parameters": [
                    {
                        "name": "pesanan",
                        "in": "path",
                        "description": "ID pesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pesanan lengkap"
                    },
                    "403": {
                        "description": "Bukan pemilik pesanan ini"
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login dulu lewat /login, lalu isi token yang didapat di sini (tanpa kata \"Bearer\").",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}